Rename module to singular
